+++ src/group.js
@@ -17,11 +17,11 @@ SVG.G = SVG.invent({
     }
     // Move by center over x-axis
   , cx: function(x) {
-      return x == null ? this.bbox().cx : this.x(x - this.bbox().width / 2)
+      return x == null ? this.tbox().cx : this.x(x - this.tbox().width / 2)
     }
     // Move by center over y-axis
   , cy: function(y) {
-      return y == null ? this.bbox().cy : this.y(y - this.bbox().height / 2)
+      return y == null ? this.tbox().cy : this.y(y - this.tbox().height / 2)
     }
   }