+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - fixed a bug where events are not detached properly -> __TODO!__
 - fixed a bug in IE11 with `mouseenter` and `mouseleave` -> __TODO!__
 
-# 1.0.0-rc.10 (?/09/2014)
+# 1.0.0-rc.10 (31/08/2014)
 
 - implemented an SVG adoption system to be able to manipulate existing SVG's not created with svg.js
 - changed `parent` reference on elements to `parent()` method
@@ -51,6 +51,8 @@
 - implemented the `element()` method using the `SVG.Bare` class to create elements that are not described by SVG.js
 - removed `SVG.Symbol` but kept the `symbol()` method using the new `element()` method
 - reworked `SVG.Number` to return new instances with calculations rather than itself
+- added `w` and `h` properties as shorthand for `width` and `height` to `SVG.BBox`
+- added `SVG.TBox` to get a bounding box that is affected by transformation values
 
 # 1.0.0-rc.9 (17/06/2014)